WIP: Update rubric submission logic to handle trying to submit a rubric that already exists
Frontend: The current state of prompting the user for their choice uses browser dialog instead of using React. This is just a prototype for implementing the underlying logic and will be refactored later.
Backend: There is currently a Prisma error when attempting to update a rubric by id.

PrismaClientKnownRequestError:
Invalid `prisma.rubric.update()` invocation:

An operation failed because it depends on one or more records that were required but not found. No 'RubricCriterion' record was found for a nested update on relation 'RubricToRubricCriterion'.